Add benchmark

+++ mxm.h
@@ -2,6 +2,7 @@
 #define MXM
 
 // Carry out a M x M matrix multiplication
-int* mxm(int*, int*, int);
+#include <stdint.h>
+uint8_t* mxm(uint8_t*, uint8_t*, uint64_t);
 
 #endif
